requirements.txt
kopf==1.29.2
kubernetes==12.0.1
loguru==0.5.3
PyYAML==5.4.1
